Treatment iteration
- fixed food reservation throwing errors
- added prisoner sucicides
- added +5 bonus to mood while prisoner have free time
- added +5 bonus to mood if prisoner is not supervised and got lazy
- added +15 bonus to mood if treatment is above 75%
- added blocking revolts (100%, 95%, 50%, 10%) if overall prisoner treatment is good enough
- increased base chance for Revolts
- added blocking mental breaks for prisoners with low treatment levels
- added treatment drop when prisoner is being beaten
- fixed cutting some content of bill config in some languages
- cosmetic changes to bill checkbox
- prisoners will now give offer to join colony if treatment is good enough (random)
- prisoners will now stay at bed if waiting for surgery
- "Treatment happiness" will decrease if health conditions are bad, when prisoners are hungry, or they're working.
- fixed blurred effect on settings window

+++ Source/SettingsMenu.cs
@@ -43,7 +43,7 @@ namespace PrisonLabor
         public override void DoSettingsWindowContents(Rect inRect)
         {
             var leftRect = new Rect(inRect.x, inRect.y, inRect.width * 0.65f, inRect.height);
-            var rightRect = new Rect(inRect.x + inRect.width * 0.65f + 30f, inRect.y, inRect.width * 0.35f - 30f,
+            var rightRect = new Rect((int)(inRect.x + inRect.width * 0.65f + 30f), inRect.y, inRect.width * 0.35f - 30f,
                 inRect.height);
 
             var listing_options = new Listing_Standard();
